LPD-51753 Create new FieldFile component
