+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/components/forms/index.ts
@@ -3,5 +3,6 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
+export {default as FieldFile} from './FieldFile';
 export {default as FieldPicker} from './FieldPicker';
 export {default as FieldText} from './FieldText';
