+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/components/modal/ImportStructureModalContent.tsx
@@ -4,11 +4,11 @@
  */
 
 import ClayAlert from '@clayui/alert';
-import ClayButton, {ClayButtonWithIcon} from '@clayui/button';
-import {ClayInput} from '@clayui/form';
+import ClayButton from '@clayui/button';
 import ClayModal from '@clayui/modal';
-import {sub} from 'frontend-js-web';
-import React, {useRef, useState} from 'react';
+import React, {useState} from 'react';
+
+import {FieldFile} from '../forms';
 
 const JSON_EXTENSION = '.json';
 
@@ -18,36 +18,6 @@ export default function ImportStructureModalContent({
 	closeModal: () => void;
 }) {
 	const [warning, setWarning] = useState(true);
-	const [jsonFile, setJsonFile] = useState<File | null>(null);
-	const fileInputRef = useRef<HTMLInputElement | null>(null);
-
-	const resetFileInput = () => {
-		if (fileInputRef && fileInputRef.current) {
-			fileInputRef.current.value = '';
-		}
-
-		setJsonFile(null);
-	};
-
-	const onChangeClick = () => {
-		fileInputRef.current?.click();
-
-		resetFileInput();
-	};
-
-	const handleFileInputChange = async ({
-		target,
-	}: React.ChangeEvent<HTMLInputElement>) => {
-		if (
-			!target.value.endsWith(JSON_EXTENSION) ||
-			!target.files ||
-			target.files?.length === 0
-		) {
-			return;
-		}
-
-		setJsonFile(target.files[0]);
-	};
 
 	return (
 		<>
@@ -71,75 +41,10 @@ export default function ImportStructureModalContent({
 			)}
 
 			<ClayModal.Body>
-				<ClayInput.Group>
-					<ClayInput.GroupItem>
-						<label htmlFor="jsonInputId">
-							{Liferay.Language.get('json-file')}
-						</label>
-
-						<ClayInput
-							id="jsonInputId"
-							value={jsonFile?.name || ''}
-						/>
-
-						<input
-							accept={JSON_EXTENSION}
-							className="d-none"
-							id="fileInputId"
-							onChange={handleFileInputChange}
-							ref={fileInputRef}
-							type="file"
-						/>
-					</ClayInput.GroupItem>
-
-					<ClayInput.GroupItem className="mt-4" shrink>
-						{jsonFile ? (
-							<>
-								<ClayButtonWithIcon
-									aria-label={sub(
-										Liferay.Language.get('change-x'),
-										Liferay.Language.get('file')
-									)}
-									className="lfr-portal-tooltip"
-									displayType="secondary"
-									onClick={onChangeClick}
-									symbol="change"
-									title={sub(
-										Liferay.Language.get('change-x'),
-										Liferay.Language.get('file')
-									)}
-									type="button"
-								/>
-
-								<ClayButtonWithIcon
-									aria-label={sub(
-										Liferay.Language.get('remove-x'),
-										Liferay.Language.get('file')
-									)}
-									className="lfr-portal-tooltip"
-									displayType="unstyled"
-									onClick={resetFileInput}
-									symbol="trash"
-									title={sub(
-										Liferay.Language.get('remove-x'),
-										Liferay.Language.get('file')
-									)}
-									type="button"
-								/>
-							</>
-						) : (
-							<ClayButtonWithIcon
-								aria-label={Liferay.Language.get('add')}
-								className="lfr-portal-tooltip"
-								displayType="secondary"
-								onClick={() => fileInputRef.current?.click()}
-								symbol="plus"
-								title={Liferay.Language.get('add')}
-								type="button"
-							/>
-						)}
-					</ClayInput.GroupItem>
-				</ClayInput.Group>
+				<FieldFile
+					label={Liferay.Language.get('json-file')}
+					validExtensions={JSON_EXTENSION}
+				/>
 			</ClayModal.Body>
 
 			<ClayModal.Footer
@@ -153,7 +58,7 @@ export default function ImportStructureModalContent({
 							{Liferay.Language.get('cancel')}
 						</ClayButton>
 
-						<ClayButton disabled={!jsonFile} displayType="primary">
+						<ClayButton displayType="primary">
 							{Liferay.Language.get('import-and-override')}
 						</ClayButton>
 					</ClayButton.Group>
